+++ DataAddons/Retail/07_Legion/ZoneData.lua
@@ -1,0 +1,11 @@
+local _, addon = ...;
+local shared = addon.Data.ZoneData.Shared;
+
+KrowiAF.ZoneData.Legion = {
+    { -- Ruins of Falanaar (zone)
+        {692, 693},
+        {
+            -- 19559, -- Azj-Kahet
+        }
+    },
+};