+++ DataAddons/Retail/10_Dragonflight/ZoneData.lua
@@ -1,0 +1,129 @@
+local _, addon = ...;
+local shared = addon.Data.ZoneData.Shared;
+
+KrowiAF.ZoneData.Dragonflight = {
+    { -- Vault of the Incarnates (raid)
+        {2119, 2120, 2121, 2122, 2123, 2124, 2125, 2126},
+        {
+            17110, -- The Primal Bulwark
+            17111, -- Caverns of Infusion
+            17112, -- Fury of the Storm
+            16343, -- Vault of the Incarnates
+            16345, -- Heroic: Vault of the Incarnates
+            16354, -- Mythic: Vault of the Incarnates
+            17107, -- Ahead of the Curve: Raszageth the Storm-Eater
+            17108, -- Cutting Edge: Raszageth the Storm-Eater
+            16395, -- Vaulternative Fashion
+            17119, -- Deep Cuts From the Vault
+            19564, -- Awakened Storms
+            19565, -- Heroic: Awakened Storms
+            19566, -- Mythic: Awakened Storms
+            16335, -- What Frozen Things Do
+            16365, -- Little Friends
+            16364, -- The Lunker Below
+            16419, -- I Was Saving That For Later
+            16458, -- Nothing But Air
+            16450, -- The Power is MINE!
+            16442, -- Incubation Extermination
+            16451, -- The Ol Raszle Daszle
+            16355, -- Glory of the Vault Raider
+            16346, -- Mythic: Eranog
+            16347, -- Mythic: Terros
+            16348, -- Mythic: The Primal Council
+            16349, -- Mythic: Sennarth, The Cold Breath
+            16350, -- Mythic: Dathea, Ascended
+            16351, -- Mythic: Kurog Grimtotem
+            16352, -- Mythic: Broodkeeper Diurna
+            16353, -- Mythic: Raszageth the Storm-Eater
+            19574, -- Awakening the Dragonflight Raids
+            19575, -- Heroic: Awakening the Dragonflight Raids
+            19576, -- Mythic: Awakening the Dragonflight Raids
+        }
+    },
+    { -- Aberrus, the Shadowed Crucible (raid)
+        {2166, 2167, 2168, 2169, 2170, 2171, 2172, 2173, 2174},
+        {
+            18163, -- Discarded Works
+            18164, -- Fury of Giants
+            18165, -- Neltharion's Shadow
+            18167, -- Edge of the Void
+            18160, -- Aberrus, the Shadowed Crucible
+            18161, -- Heroic: Aberrus, the Shadowed Crucible
+            18162, -- Mythic: Aberrus, the Shadowed Crucible
+            18253, -- Ahead of the Curve: Scalecommander Sarkareth
+            18254, -- Cutting Edge: Scalecommander Sarkareth
+            17765, -- What We Wear In The Shadowflame
+            18256, -- Nasz'uro, the Unbound Legacy
+            18027, -- Dragonflight Season 2 Master
+            18380, -- Dragonflight Season 2 Hero
+            19567, -- Awakened Shadows
+            19568, -- Heroic: Awakened Shadows
+            19569, -- Mythic: Awakened Shadows
+            18229, -- Cosplate
+            18168, -- I'll Make My Own Shadowflame
+            18173, -- Tabula Rasa
+            18228, -- Are You Even Trying?
+            18230, -- Whac-A-Swog
+            18193, -- Eggscellent Eggsecution
+            18172, -- Escar-Go-Go-Go
+            18149, -- Objects in Transit May Shatter
+            17877, -- We'll Never See That Again, Surely
+            18251, -- Glory of the Aberrus Raider
+            18151, -- Mythic: Kazzara, the Hellforged
+            18152, -- Mythic: The Amalgamation Chamber
+            18153, -- Mythic: The Forgotten Experiments
+            18154, -- Mythic: Assault of the Zaqali
+            18155, -- Mythic: Rashok, the Elder
+            18156, -- Mythic: The Vigilant Steward, Zskarn
+            18157, -- Mythic: Magmorax
+            18158, -- Mythic: Echo of Neltharion
+            18159, -- Mythic: Scalecommander Sarkareth
+            19574, -- Awakening the Dragonflight Raids
+            19575, -- Heroic: Awakening the Dragonflight Raids
+            19576, -- Mythic: Awakening the Dragonflight Raids
+        }
+    },
+    { -- Amirdrassil, the Dream's Hope (raid)
+        {2232, 2233, 2234, 2235, 2236, 2237, 2238, 2240, 2244},
+        {
+            19344, -- Incarnate's Wake
+            19345, -- Molten Incursion
+            19346, -- The Viridian Weave
+            19347, -- Fate of Amirdrassil
+            19331, -- Amirdrassil, the Dream's Hope
+            19333, -- Heroic: Amirdrassil, the Dream's Hope
+            19334, -- Mythic: Amirdrassil, the Dream's Hope
+            19350, -- Ahead of the Curve: Fyrakk the Blazing
+            19351, -- Cutting Edge: Fyrakk the Blazing
+            19442, -- Fire Catwalk With Me
+            19450, -- Fyr'alath the Dreamrender
+            19396, -- Dragonflight Season 3 Master
+            19420, -- Dragonflight Season 3 Hero
+            19570, -- Awakened Flames
+            19571, -- Heroic: Awakened Flames
+            19572, -- Mythic: Awakened Flames
+            19322, -- Meaner Pastures
+            19320, -- Cruelty Free
+            19321, -- Swog Champion
+            19193, -- Ducks In A Row
+            19089, -- Don't Let the Doe Hit You On The Way Out
+            19394, -- A Dream Within a Dream
+            19319, -- Haven't We Done This Before?
+            19393, -- Whelp, I'm Lost
+            19390, -- Memories of Teldrassil
+            19349, -- Glory of the Dream Raider
+            19335, -- Mythic: Gnarlroot
+            19336, -- Mythic: Igira the Cruel
+            19337, -- Mythic: Volcoross
+            19338, -- Mythic: Council of Dreams
+            19339, -- Mythic: Larodar, Keeper of the Flame
+            19340, -- Mythic: Nymue, Weaver of the Cycle
+            19341, -- Mythic: Smolderon
+            19342, -- Mythic: Tindral Sageswift, Seer of the Flame
+            19343, -- Mythic: Fyrakk the Blazing
+            19574, -- Awakening the Dragonflight Raids
+            19575, -- Heroic: Awakening the Dragonflight Raids
+            19576, -- Mythic: Awakening the Dragonflight Raids
+        }
+    },
+};