+++ DataAddons/Retail/11_TheWarWithin/TransmogSetData.lua
@@ -33,4 +33,66 @@ KrowiAF.TransmogSetData.TheWarWithin = {
             3829, 3842, 3821, 3834, 3833, 3846, -- Forged Gladiator's Silk Armor (Gladiator, Elite)
         }
     },
+    { -- Can You Please Spell "Gobanna?"
+        41525,
+        {
+            4326, 4325, 4323, 4324, -- Enforcer's Backalley Brawlplate (Raid Finder, Normal, Heroic, Mythic)
+            4306, 4305, 4303, 4304, -- Oath of the Aureate Sentry (Raid Finder, Normal, Heroic, Mythic)
+            4294, 4293, 4291, 4292, -- Tireless Collector's Bounties (Raid Finder, Normal, Heroic, Mythic)
+            4314, 4313, 4311, 4312, -- Spectral Gambler's Last Call (Raid Finder, Normal, Heroic, Mythic)
+            4310, 4309, 4307, 4308, -- Confessor's Unshakable Virtue (Raid Finder, Normal, Heroic, Mythic)
+            4278, 4277, 4275, 4276, -- Cauldron Champion's Encore (Raid Finder, Normal, Heroic, Mythic)
+            4318, 4317, 4315, 4316, -- Currents of the Gale Sovereign (Raid Finder, Normal, Heroic, Mythic)
+            4298, 4297, 4295, 4296, -- Jewels of the Aspectral Emissary (Raid Finder, Normal, Heroic, Mythic)
+            4322, 4321, 4319, 4320, -- Spliced Fiendtrader's Influence (Raid Finder, Normal, Heroic, Mythic)
+            4302, 4301, 4299, 4300, -- Ageless Serpent's Foresight (Raid Finder, Normal, Heroic, Mythic)
+            4286, 4285, 4283, 4284, -- Roots of Reclaiming Blight (Raid Finder, Normal, Heroic, Mythic)
+            4282, 4281, 4279, 4280, -- Fel-Dealer's Contraband (Raid Finder, Normal, Heroic, Mythic)
+            4290, 4289, 4287, 4288, -- Opulent Treasurescale's Hoard (Raid Finder, Normal, Heroic, Mythic)
+        }
+    },
+    { -- Prized Guise
+        41595,
+        {
+            4340, -- Prized Aspirant's Plate Armor (Aspirant)
+            4017, 4031, 4013, 4026, 4006, 4019, -- Prized Gladiator's Plate Armor (Gladiator, Elite)
+            4339, -- Prized Aspirant's Chain Armor (Aspirant)
+            4010, 4023, 4016, 4029, 4009, 4022, -- Prized Gladiator's Chain Armor (Gladiator, Elite)
+            4338, -- Prized Aspirant's Leather Armor (Aspirant)
+            4015, 4028, 4012, 4025, 4008, 4021, 4007, 4020, -- Prized Gladiator's Leather Armor (Gladiator, Elite)
+            4337, -- Prized Aspirant's Silk Armor (Aspirant)
+            4014, 4027, 4011, 4024, 4018, 4030, -- Prized Gladiator's Silk Armor (Gladiator, Elite)
+        }
+    },
+    { -- Fractals? For Spring? Groundbreaking.
+        42316,
+        {
+            5148, 5147, 5145, 5146, -- Chains of the Living Weapon (Raid Finder, Normal, Heroic, Mythic)
+            5128, 5127, 5125, 5126, -- Vows of the Lucent Battalion (Raid Finder, Normal, Heroic, Mythic)
+            5116, 5115, 5113, 5114, -- Midnight Herald's Pledge (Raid Finder, Normal, Heroic, Mythic)
+            5136, 5135, 5133, 5134, -- Shroud of the Sudden Eclipse (Raid Finder, Normal, Heroic, Mythic)
+            5132, 5131, 5129, 5130, -- Eulogy to a Dying Star (Raid Finder, Normal, Heroic, Mythic)
+            5100, 5099, 5097, 5098, -- Hollow Sentinel's Vigil (Raid Finder, Normal, Heroic, Mythic)
+            5140, 5139, 5137, 5138, -- Howls of Channeled Fury (Raid Finder, Normal, Heroic, Mythic)
+            5120, 5119, 5117, 5118, -- Augur's Ephemeral Plumage (Raid Finder, Normal, Heroic, Mythic)
+            5144, 5143, 5141, 5142, -- Inquisitor's Feast of Madness (Raid Finder, Normal, Heroic, Mythic)
+            5124, 5123, 5121, 5122, -- Crash of Fallen Storms (Raid Finder, Normal, Heroic, Mythic)
+            5108, 5107, 5105, 5106, -- Ornaments of the Mother Eagle (Raid Finder, Normal, Heroic, Mythic)
+            5104, 5103, 5101, 5102, -- Charhound's Vicious Hunt (Raid Finder, Normal, Heroic, Mythic)
+            5112, 5111, 5109, 5110, -- Spellweaver's Immaculate Design (Raid Finder, Normal, Heroic, Mythic)
+        }
+    },
+    { -- Astral Attire
+        42800,
+        {
+            5170, -- Astral Aspirant's Plate Armor (Aspirant)
+            4110, 4124, 4106, 4119, 4099, 4112, -- Astral Gladiator's Plate Armor (Gladiator, Elite)
+            5171, -- Astral Aspirant's Chain Armor (Aspirant)
+            4103, 4116, 4109, 4122, 4102, 4115, -- Astral Gladiator's Chain Armor (Gladiator, Elite)
+            5172, -- Astral Aspirant's Leather Armor (Aspirant)
+            4108, 4121, 4105, 4118, 4101, 4114, 4100, 4113, -- Astral Gladiator's Leather Armor (Gladiator, Elite)
+            5173, -- Astral Aspirant's Silk Armor (Aspirant)
+            4107, 4120, 4104, 4117, 4111, 4123, -- Astral Gladiator's Silk Armor (Gladiator, Elite)
+        }
+    },
 };