+++ Localization/enUS.lua
@@ -7,7 +7,7 @@ KrowiAF.PluginsApi:LoadPluginLocalization(L);
 -- [[ https://legacy.curseforge.com/wow/addons/krowi-achievement-filter/localization ]] --
 -- [[ Everything after this line is automatically generated from CurseForge and is not meant for manual edit - SOURCETOKEN - AUTOGENTOKEN ]] --
 
--- [[ Exported at 2025-06-18 15-01-34 ]] --
+-- [[ Exported at 2025-10-01 21-40-53 ]] --
 L["%c"] = true
 L["%d/%m/%Y %I:%M %p"] = true
 L["%m/%d/%Y %I:%M %p"] = true
@@ -47,8 +47,8 @@ This value can be changed while the achievement window is open and having one of
 (1) FPS drops and stuttering while adjusting the value, NOT during normal addon use]=]
 L["Active"] = true
 L["Active Event Login Delay"] = true
-L["Active events"] = true
 L["Active Events"] = true
+L["Active events"] = true
 L["Add addon name to world map icon"] = true
 L["Add addon name to world map icon Desc"] = "Add '{addonName}' to the bottom of the tooltip when mousing over the icon."
 L["Add Locale"] = true
@@ -138,8 +138,8 @@ This value can be changed while the achievement window is open and having one of
 (1) FPS drops and stuttering while adjusting the value, NOT during normal addon use]=]
 L["Category shown temporarily"] = "This category was selected by clicking on the {eventName} event reminder and is shown temporarily because it has no achievements that match the set filters."
 L["Caverns Criss-Cross"] = true
-L["Center header"] = true
 L["Center Header"] = true
+L["Center header"] = true
 L["Center header Desc"] = "Keep the header centered on the achievements window."
 L["Center Header Desc"] = "Center the achievement's name on the achievement button."
 L["Central Circuit"] = true
@@ -193,8 +193,8 @@ L["Completion"] = true
 L["Copy Account Wide to Character"] = true
 L["Copy Account Wide to Character Desc"] = "Copy the account wide {watchList} (any achievement watched by any character) to the current character specific {watchList}. From now on this character will have it's own {watchList}."
 L["Covenant"] = true
-L["Covenant assaults"] = true
 L["Covenant Assaults"] = true
+L["Covenant assaults"] = true
 L["Credits"] = true
 L["Criteria"] = true
 L["Criteria Behaviour"] = true
@@ -578,6 +578,7 @@ L["Left Ctrl"] = true
 L["Left Shift"] = true
 L["Legacy of Arathor"] = true
 L["Legion Assaults"] = true
+L["Legion: Remix"] = true
 L["Level 80"] = true
 L["Level 85"] = true
 L["Level 90"] = true
@@ -781,7 +782,8 @@ L["Related Tab"] = true
 L["Related Tab Desc"] = "Adds an extra bit to the {wowheadLink} so another tab different than the default one will be selected."
 L["Remember frame position"] = "Remember {frame} position"
 L["Remember frame position Desc"] = "Remember the {frame} position. This will keep the {frame} in it's last known position the next time you log in."
-L["Remix Pandaria Bronze"] = true
+L["Remix Bronze"] = true
+L["Remix Infinite Knowledge"] = true
 L["Remove from Watch List"] = "Remove from {watchList}"
 L["Required for"] = true
 L["Requires"] = true